Tamper Detection from RPC Working
The Diamond-HSM software can now send a RPC to the CrypTech
devices to check for a tamper event. The RPC function
needs to be updated in the STM32 code and currently will
report a tamper event after the first 60 seconds.

Requirements to build firmware
CrypTech/sw/libhal - featureDiamondKeyTamper c565518e32
CrypTech/sw/stm32  - featureTamperFirmware   b641ba0e65

The correct firmware has been commited to this branch

Another cool feature is the HSM now tracks updates to the
STM and Tiny firmware. When it detects that the firmware
is out of date, it will restrict access until the firmware
has been updated.

The HSM can now also track 'dirty' restarts, restarts
that happened because of a powerloss and not because the
HSM was restarted because of a CTY command.